ble-hid works, but I'm gonna have to undo a lot of hacks
Main issue took a while to figure out: CONFIG_SETTINGS (&
CONFIG_BT_SETTINGS) also use the `storage` partition. Since we use it at
the same time for LittleFS, shit goes down and it's really hard to
debug. Just a bunch of gibberish in logs, which I thought it was a stack
overflow and chased that wrong thread for hours...